Fix json encoding
- Changing atoi to strtoull and strtoll as atoi return only 32 bit
  values. strtoll and strtoull return 64bit values which we encode our
inetegers as.
- Changing a few other places where json encoding should use 64 bit
  values instead of 32 bit.

+++ sys/log/src/log_nmgr.c
@@ -58,7 +58,7 @@ static struct nmgr_handler log_nmgr_group_handlers[] = {
 struct encode_off {
     struct json_encoder *eo_encoder;
     int64_t eo_ts;
-    uint32_t eo_index;
+    uint8_t eo_index;
 };
 
 static int
